added flag-based xml2png.py
